RAG/requirements.txt
torch>=1.9.0
transformers>=4.21.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
chromadb>=0.4.0
langchain>=0.0.200
openai>=0.27.0
tiktoken>=0.4.0
numpy>=1.21.0
pandas>=1.3.0
python-dotenv>=1.0.0
streamlit>=1.25.0